Fixes issue with copy module when inherit permissions is unchecked
When a user with edit rights (aka admin, NOT host) copies an existing module that has "inherit permissions" unchecked, no permissions are saved for the new module. 

For the admin user the module "vanishes" from the page (seems to be not added). However the host user still sees the module.



fix issue #5633