Updated Version to handle Azure SQL
- Fixes #6499

Azure SQL Server is reported using a 12 version number, but  it is ALWAYS updated to the latest and greatest version so it meets the current 2017 and later requirement that is needed by DNN Platform.

This change adds a special secondary check if a 12 version number is returned to check for the edition.  If it is Azure SQL it sets the proper flag allowing installation