Fixed an issue where CustumCss module would not load
Closes #6601

The global personaBarSettings javascript objet is merged from the backend with some portal settings, I guess in some place, stuff was changed from PortalSettings to IPortalSettings which means we now have `PortalId` instead of `PortalID`.

Luckily, it appears that value was only used in a few places, so this PR changes the casing in those places fixing the CSS module and some more areas affected by the same issue.