Fixed Cookie Timeout Lookup Values
The existing behavior blindly assumed that if a location node exists that the system.web.authencation/forms element is in the location node.

This updates the behavior to start with the normal location, IF if funds forms authenticaton it continues, otherwise, it tries to find it under a location node IF location nodes exist.

This should restore functionality after the updates recently resulted in ALL installations having a location node.  This has been tested and appears to be working well